backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
websockets
python-multipart
elevenlabs
httpx
pydantic
sqlalchemy
bcrypt
pyjwt
